fixed some inheritance issues in projectors
