+++ neuralxc/projector/__init__.py
@@ -1,6 +1,6 @@
 # from .projector import DensityProjector,BehlerProjector, NonOrthoProjector, DeltaProjector, DefaultProjector, BaseProjector, M_make_complex
 from .projector import (BaseProjector, DensityProjector, EuclideanProjector, RadialProjector)
-from .gaussian import GaussianProjector, GaussianRadialProjector
-from .polynomial import OrthoProjector, OrthoRadialProjector
+from .gaussian import GaussianEuclideanProjector, GaussianRadialProjector
+from .polynomial import OrthoEuclideanProjector, OrthoRadialProjector
 from .pyscf import PySCFProjector
 from . import gaussian, polynomial, projector
