+++ neuralxc/projector/projector.py
@@ -79,32 +79,6 @@ class BaseProjector(TorchModule, metaclass=ProjectorRegistry):
     def set_species(self, species):
         self.species = species
 
-    def forward_basis(self, positions, unitcell, grid, my_box):
-        """Creates basis set (for projection) for a single atom, on grid points
-
-        Parameters
-        ----------
-        positions, Tensor (1, 3) or (3)
-        	atomic position
-        unitcell, Tensor (3,3)
-        	Unitcell in bohr
-        grid, Tensor (3)
-        	Grid points per unitcell
-        my_box, Tensor (3: euclid. directions, 2: upper and lower limits)
-            Limiting box local gridpoints. Relevant if global grid is decomposed
-            with MPI or similar.
-
-        Returns
-        --------
-        rad, ang, mesh
-            Stacked radial and angular functions as well as meshgrid
-        """
-
-        self.set_cell_parameters(unitcell, grid)
-        basis = self.basis[self.species]
-        box, mesh = self.box_around(positions, basis['r_o'], my_box)
-        rad, ang = self.get_basis_on_mesh(box, basis, self.W[self.species])
-        return rad, ang, mesh
 
     def forward(self, rho, positions, species, unitcell, grid, my_box):
         """ Combines basis set creation (done in forward_basis) and projection
@@ -172,34 +146,6 @@ class BaseProjector(TorchModule, metaclass=ProjectorRegistry):
         for m in range(-l, l + 1):
             res.append(geom.SH(l, m, theta, phi))
         return res
-
-    def project_onto(self, rho, rads, angs, n_l):
-        rho = rho.squeeze()
-        rho = rho * self.V_cell.squeeze()
-        print(rho.size())
-        if rho.ndim == 1:
-            coeff_array = contract('li,ni,i -> nl', angs, rads, rho)
-        else:
-            coeff_array = contract('lmijk,nijk,ijk -> nlm', angs, rads, rho)
-
-        return coeff_array.view(-1)
-
-    def get_basis_on_mesh(self, box, basis, W):
-
-        n_rad = basis['n']
-        n_l = basis['l']
-        r_o = basis['r_o']
-        R, Theta, Phi = box['radial']
-
-        #Build angular part of basis functions
-        angs = []
-        for l in range(n_l):
-            angs += self.angulars_real(l, Theta, Phi)
-
-        angs = torch.stack(angs)
-        rads = self.radials(R, basis, W)
-
-        return rads, angs
 
 
 class EuclideanProjector(BaseProjector):
@@ -365,7 +311,7 @@ class EuclideanProjector(BaseProjector):
             return Rm
 
 
-class RadialProjector(EuclideanProjector):
+class RadialProjector(BaseProjector):
 
     _registry_name = 'radial'
 
@@ -385,7 +331,7 @@ class RadialProjector(EuclideanProjector):
         basis_instructions: dict
         	Instructions that defines basis
         """
-        TorchModule.__init__(self)
+        BaseProjector.__init__(self)
         self.basis = basis_instructions
         # Initialize the matrix used to orthonormalize radial basis
         W = {}
